% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_add_page.R
\name{add_page}
\alias{add_page}
\alias{add_page.pivot_table}
\title{Title}
\usage{
add_page(pt, row = 0, col = 0, page = "")

\method{add_page}{pivot_table}(pt, row = 0, col = 0, page = "")
}
\arguments{
\item{page}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
