% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{assign_names}
\alias{assign_names}
\title{data frame col and row names}
\usage{
assign_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame.
}
\description{
Assign names to the rows and columns of the data frame.
}
\keyword{internal}
