% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_change_dec.R
\name{change_dec}
\alias{change_dec}
\alias{change_dec.pivot_table}
\title{Title}
\usage{
change_dec(pt, sep = ".")

\method{change_dec}{pivot_table}(pt, sep = ".")
}
\arguments{
\item{sep}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
