% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_define_labels.R
\name{define_labels}
\alias{define_labels}
\alias{define_labels.pivot_table}
\title{Title}
\usage{
define_labels(pt, n_col, n_row)

\method{define_labels}{pivot_table}(pt, n_col, n_row)
}
\arguments{
\item{n_row}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
