% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ex}
\alias{df_ex}
\title{Pivot table in data frame with with thousands indicator and decimal numbers}
\format{
A data frame.
}
\usage{
df_ex
}
\description{
Pivot table in data frame with with thousands indicator and decimal numbers.
}
\seealso{
\code{\link{pt_ex}}

Other pivot table in data frame: 
\code{\link{df_ex_compact}},
\code{\link{df_pivottabler}}
}
\concept{pivot table in data frame}
\keyword{datasets}
