% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ex_compact}
\alias{df_ex_compact}
\title{Pivot table in data frame with a column with data from two label fields}
\format{
A data frame.
}
\usage{
df_ex_compact
}
\description{
Pivot table in data frame in compact table format: with a column with data
from two label fields.
}
\seealso{
\code{\link{pf_ex_compact}}

Other pivot table in data frame: 
\code{\link{df_ex}},
\code{\link{df_pivottabler}}
}
\concept{pivot table in data frame}
\keyword{datasets}
