% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_pivottabler}
\alias{df_pivottabler}
\title{Pivot table with basic and subtotal labels in the same column}
\format{
A data frame.
}
\source{
\url{https://CRAN.R-project.org/package=pivottabler}
}
\usage{
df_pivottabler
}
\description{
A dataset containing number of train passengers, generated with the
\code{pivottabler} package. It contains basic and subtotal labels in the same column.
}
\seealso{
Other pivot table in data frame: 
\code{\link{df_ex_compact}},
\code{\link{df_ex}}
}
\concept{pivot table in data frame}
\keyword{datasets}
