% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_set_v}
\alias{df_set_v}
\title{Set of pivot tables placed vertically on one sheet}
\format{
A data frame.
}
\usage{
df_set_v
}
\description{
Set of pivot tables placed vertically on one sheet.
}
\keyword{datasets}
