% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_divide.R
\name{divide}
\alias{divide}
\alias{divide.pivot_table}
\title{Title}
\usage{
divide(pt)

\method{divide}{pivot_table}(pt)
}
\arguments{
\item{pt}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
