% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_col_values.R
\name{extract_col_values}
\alias{extract_col_values}
\title{Title}
\usage{
extract_col_values(lpt, col = 1, start_row = 2, rows_left = 0)
}
\arguments{
\item{rows_left}{}
}
\value{

}
\description{
Title
}
