% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_extract_label.R
\name{extract_label}
\alias{extract_label}
\alias{extract_label.pivot_table}
\title{Title}
\usage{
extract_label(pt, col = 1, values = c())

\method{extract_label}{pivot_table}(pt, col = 1, values = c())
}
\arguments{
\item{values}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
