% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_table.R
\name{flatten_table}
\alias{flatten_table}
\title{Title}
\usage{
flatten_table(
  pt,
  add_page_row = 0,
  add_page_col = 0,
  rem_top = 0,
  rem_bottom = 0,
  rem_left = 0,
  rem_right = 0,
  n_col_labels = 1,
  n_row_labels = 1,
  extract_label_col = 0,
  extract_label_values = c(),
  rem_k = ".",
  rem_agg = NULL,
  include_page = TRUE,
  include_na = FALSE
)
}
\arguments{
\item{include_na}{}
}
\value{

}
\description{
Title
}
