% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_ex_v2}
\alias{ft_ex_v2}
\title{Flat table without page from a pivot table with with thousands indicator and decimal numbers}
\format{
A \code{tibble} object.
}
\usage{
ft_ex_v2
}
\description{
Flat table without page from a pivot table with with thousands indicator and decimal numbers.
}
\examples{
# Defined by:

ft_ex_v2 <- pivot_table(df_ex) |>
  set_page(1, 1) |>
  remove_top(1) |>
  define_labels(n_col = 2, n_row = 2) |>
  remove_k() |>
  replace_dec() |>
  fill_values() |>
  fill_labels() |>
  remove_agg() |>
  unpivot(include_page = FALSE,
  na_rm = FALSE)

}
\seealso{
\code{\link{df_ex}}

Other flat table: 
\code{\link{ft_ex}},
\code{\link{ft_set}}
}
\concept{flat table}
\keyword{datasets}
