% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ft_set}
\alias{ft_set}
\title{Flat table with page from a pivot table with with thousands indicator and decimal numbers}
\format{
A \code{tibble} object.
}
\usage{
ft_set
}
\description{
Flat table with page from a pivot table with with thousands indicator and decimal numbers.
}
\examples{
# Defined by:

f <- function(pt) {
 pt |>
    set_page(1, 1) |>
    remove_top(1) |>
    define_labels(n_col = 2, n_row = 2) |>
    remove_k() |>
    replace_dec() |>
    fill_values() |>
    fill_labels() |>
    remove_agg() |>
    unpivot()
}

pt <- pivot_table(df_set_h_v)
lpt <- pt |> divide()
ft_set <- flatten_table_list(lpt, f)

}
\seealso{
\code{\link{df_set_h_v}}

Other flat table: 
\code{\link{ft_ex_v2}},
\code{\link{ft_ex}}
}
\concept{flat table}
\keyword{datasets}
