% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_pt_compact}
\alias{list_pt_compact}
\title{Pivot tables with a column with data from two label fields}
\format{
List of \code{pivot_table} objects.
}
\usage{
list_pt_compact
}
\description{
List of pivot tables with a column with data from two label fields (this is
generally known as compact table format).
}
\keyword{datasets}
