% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_pt_ie}
\alias{list_pt_ie}
\title{List of pivot tables used in the vignette}
\format{
List of \code{pivot_table} objects.
}
\usage{
list_pt_ie
}
\description{
List of pivot tables used in the vignette, an illustrative example.
}
\keyword{datasets}
