% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{new_pivot_table}
\alias{new_pivot_table}
\title{Title
low-level constructor that efficiently creates new objects with the correct structure}
\usage{
new_pivot_table(
  df = data.frame(),
  page = vector("character"),
  n_col_labels = 0,
  n_row_labels = 0
)
}
\arguments{
\item{n_row_labels}{}
}
\value{

}
\description{
Title
low-level constructor that efficiently creates new objects with the correct structure
}
