% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pf_ex_compact}
\alias{pf_ex_compact}
\title{Pivot table result of transforming a data frame with a column with data from two label fields}
\format{
A\code{pivot_table} object.
}
\usage{
pf_ex_compact
}
\description{
Pivot table result of transforming a data frame in compact table format: with a column with data
from two label fields.
}
\examples{
# Defined by:

pf_ex_compact <- pivot_table(df_ex_compact) |>
  extract_labels(col = 1,
                 labels = c("b1", "b2", "b3", "b4", "Total general"))

}
\seealso{
\code{\link{df_ex_compact}}

Other pivot table: 
\code{\link{pt_ex}}
}
\concept{pivot table}
\keyword{datasets}
