% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{\code{pivot_table} S3 class}
\usage{
pivot_table(
  df,
  page = vector("character"),
  n_col_labels = 0,
  n_row_labels = 0,
  n_extract = 0
)
}
\arguments{
\item{df}{A data frame, contains one or more pivot tables.}

\item{page}{A string, additional information associated with the pivot table.}

\item{n_col_labels}{A number, number of columns containing pivot table labels.}

\item{n_row_labels}{A number, number of rows containing pivot table labels.}

\item{n_extract}{A number, number of new columns added to the table.}
}
\value{
A \code{pivot_table} object.
}
\description{
Creates a \code{pivot_table} object from a data frame. Additional information
associated with the pivot table can be indicated. The data frame data is
converted to character data type.
}
\examples{

df <- data.frame(unclass(pt_m4)[c(1:7)])
pt <- pivot_table(df)

pt <- pivot_table(df, page = "M4")

}
\seealso{
\code{\link{pivot_table}}
}
\concept{import data functions}
