% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_ex}
\alias{pt_ex}
\title{Pivot table with with thousands indicator and decimal numbers}
\format{
A \code{pivot_table} object.
}
\usage{
pt_ex
}
\description{
Pivot table with with thousands indicator and decimal numbers.
}
\examples{
# Defined by:

pt_ex <- pivot_table(df_ex)

}
\seealso{
\code{\link{df_ex}}

Other pivot table: 
\code{\link{pf_ex_compact}}
}
\concept{pivot table}
\keyword{datasets}
