% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_ine2871}
\alias{pt_ine2871}
\title{Official population in the municipalities of Granada (Spain)}
\format{
A pivot_table object.
}
\source{
\url{https://www.ine.es/dynt3/inebase/index.htm?padre=525}
}
\usage{
pt_ine2871
}
\description{
A dataset containing official population figures resulting from the revision
of the Municipal Register on January 1 for the municipalities of Granada
(Spain). It is a pivot table that has rows before and after it.
}
\keyword{datasets}
