% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_m4}
\alias{pt_m4}
\title{Pivot table with with thousands indicator and decimal numbers}
\format{
A \code{pivot_table} object.
}
\usage{
pt_m4
}
\description{
Pivot table with with thousands indicator and decimal numbers.
}
\keyword{datasets}
