% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_m4_compact}
\alias{pt_m4_compact}
\title{Pivot table with a column with data from two label fields}
\format{
A pivot_table object.
}
\usage{
pt_m4_compact
}
\description{
Pivot table in compact table format: with a column with data from two label
fields.
}
\keyword{datasets}
