% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_set_h}
\alias{pt_set_h}
\title{Set of pivot tables placed horizontally on one sheet}
\format{
A \code{pivot_table} object.
}
\usage{
pt_set_h
}
\description{
Set of pivot tables placed horizontally on one sheet.
}
\keyword{datasets}
