% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_set_h_v}
\alias{pt_set_h_v}
\title{Set of pivot tables on one sheet}
\format{
A pivot_table object.
}
\usage{
pt_set_h_v
}
\description{
Example of a set of pivot tables located horizontally and vertically on one
sheet.
}
\keyword{datasets}
