% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_text_file.R
\name{read_text_file}
\alias{read_text_file}
\title{Import text data file}
\usage{
read_text_file(file, sep = ";")
}
\arguments{
\item{file}{UTF-8 encoded text file whose data is to be read. Each line in
the file corresponds to a row in a table.}

\item{sep}{Column separator character. Within each row, columns are separated
by this character.}
}
\value{
A two elements list of metadata and character data frame containing the data in the file.
}
\description{
Reads a UTF-8 text file with column separators and creates a list of metadata and a data frame with data from
it.
}
\seealso{
\code{\link[utils]{View}}
}
