% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_bottom}
\alias{remove_bottom}
\alias{remove_bottom.pivot_table}
\title{Title}
\usage{
remove_bottom(pt, n)

\method{remove_bottom}{pivot_table}(pt, n)
}
\arguments{
\item{n}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
