% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_cols}
\alias{remove_cols}
\alias{remove_cols.pivot_table}
\title{Title}
\usage{
remove_cols(pt, c)

\method{remove_cols}{pivot_table}(pt, c)
}
\arguments{
\item{c}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
