% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_empty}
\alias{remove_empty}
\alias{remove_empty.pivot_table}
\title{Title}
\usage{
remove_empty(pt)

\method{remove_empty}{pivot_table}(pt)
}
\arguments{
\item{pt}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
