% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove_k.R
\name{remove_k}
\alias{remove_k}
\alias{remove_k.pivot_table}
\title{Title}
\usage{
remove_k(pt, sep = ".")

\method{remove_k}{pivot_table}(pt, sep = ".")
}
\arguments{
\item{sep}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
