% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_rows}
\alias{remove_rows}
\alias{remove_rows.pivot_table}
\title{Title}
\usage{
remove_rows(pt, r)

\method{remove_rows}{pivot_table}(pt, r)
}
\arguments{
\item{r}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
