% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_divide.R
\name{spacer_columns}
\alias{spacer_columns}
\title{Spacer columns}
\usage{
spacer_columns(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A vector of numbers.
}
\description{
Gets the empty column numbers for a table. If there are several consecutive
empty columns, only one is considered. The first and last columns in the
table are also added to the list even if they are not empty.
}
