% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_divide.R
\name{spacer_rows}
\alias{spacer_rows}
\title{Spacer rows}
\usage{
spacer_rows(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A vector of numbers.
}
\description{
Gets the empty row numbers for a table. If there are several consecutive
empty rows, only one is considered. The first and last rows in the table are
also added to the list even if they are not empty.
}
