% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_unpivot.R
\name{unpivot}
\alias{unpivot}
\alias{unpivot.pivot_table}
\title{Title}
\usage{
unpivot(pt, include_page = TRUE, include_na = FALSE)

\method{unpivot}{pivot_table}(pt, include_page = TRUE, include_na = FALSE)
}
\arguments{
\item{include_na}{}
}
\value{



}
\description{
Title

Title
}
\keyword{internal}
