% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_view_table_attr.R
\name{view_table_attr}
\alias{view_table_attr}
\alias{view_table_attr.pivot_table}
\title{View table and attributes}
\usage{
view_table_attr(pt)

\method{view_table_attr}{pivot_table}(pt)
}
\arguments{
\item{pt}{A \code{pivot_table} object.}
}
\value{
A \code{pivot_table} object.
}
\description{
Displays the table and attributes of the object.
}
\seealso{


Other pivot table definition functions: 
\code{\link{define_labels}()},
\code{\link{divide}()},
\code{\link{get_page}()},
\code{\link{remove_bottom}()},
\code{\link{remove_cols}()},
\code{\link{remove_empty}()},
\code{\link{remove_left}()},
\code{\link{remove_right}()},
\code{\link{remove_rows}()},
\code{\link{remove_top}()},
\code{\link{set_page}()}
}
\concept{pivot table definition functions}
