% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_set_h_v}
\alias{df_set_h_v}
\title{Set of pivot tables on one sheet}
\format{
A data frame.
}
\usage{
df_set_h_v
}
\description{
Example of a set of pivot tables located horizontally and vertically on one
sheet.
}
\seealso{
\code{\link{df_ex}}

Other pivot table set in data frame: 
\code{\link{df_set_h}},
\code{\link{df_set_v}}
}
\concept{pivot table set in data frame}
\keyword{datasets}
