% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_fill_labels.R
\name{fill_vector}
\alias{fill_vector}
\title{Fill in missing values in a vector}
\usage{
fill_vector(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A vector.
}
\description{
Fills missing values in a vector with previous value.
}
\keyword{internal}
