% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table.R
\name{pivot_table}
\alias{pivot_table}
\title{Title
helper that provides a convenient way for others to create objects of your class
you can skip the helper if the class is for internal use only
If you expect users to also create objects, you should create a friendly helper function}
\usage{
pivot_table(df, page = vector("character"), n_col_labels, n_row_labels)
}
\arguments{
\item{n_row_labels}{}
}
\value{

}
\description{
Title
helper that provides a convenient way for others to create objects of your class
you can skip the helper if the class is for internal use only
If you expect users to also create objects, you should create a friendly helper function
}
