% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pt_pivottabler}
\alias{pt_pivottabler}
\title{Pivot table with basic and subtotal labels in the same column}
\format{
A pivot_table object.
}
\source{
\url{https://CRAN.R-project.org/package=pivottabler}
}
\usage{
pt_pivottabler
}
\description{
A dataset containing number of train passengers, generated with the
pivottabler package. It contains basic and subtotal labels in the same
column.
}
\keyword{datasets}
