% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_left}
\alias{remove_left}
\alias{remove_left.pivot_table}
\title{Remove left columns from a pivot table}
\usage{
remove_left(pt, n)

\method{remove_left}{pivot_table}(pt, n)
}
\arguments{
\item{pt}{A pivot_table object.}

\item{n}{A number, number of columns to remove.}
}
\value{
A pivot_table object.
}
\description{
Remove left columns from the pivot table represented by the object.
}
\details{
A pivot table should only contain label rows and columns, and an array of
values, usually numeric data.

All columns not belonging to the pivot table must be removed.
}
\examples{
library(tidyr)

pt <- pt_m4 \%>\% remove_left(2)

}
\keyword{internal}
