% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_remove.R
\name{remove_left}
\alias{remove_left}
\alias{remove_left.pivot_table}
\title{Remove left columns from a pivot table}
\usage{
remove_left(pt, n)

\method{remove_left}{pivot_table}(pt, n)
}
\arguments{
\item{pt}{A \code{pivot_table} object.}

\item{n}{A number, number of columns to remove.}
}
\value{
A \code{pivot_table} object.
}
\description{
Remove left columns from the pivot table represented by the object.
}
\details{
A pivot table should only contain label rows and columns, and an array of
values, usually numeric data.

All columns not belonging to the pivot table must be removed.
}
\examples{
library(tidyr)

pt <- pt_m4 \%>\% remove_left(2)

}
\seealso{


Other pivot table definition functions: 
\code{\link{define_labels}()},
\code{\link{divide}()},
\code{\link{get_page}()},
\code{\link{remove_bottom}()},
\code{\link{remove_cols}()},
\code{\link{remove_empty}()},
\code{\link{remove_right}()},
\code{\link{remove_rows}()},
\code{\link{remove_top}()},
\code{\link{set_page}()},
\code{\link{view_table_attr}()}
}
\concept{pivot table definition functions}
