% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_page.R
\name{set_page}
\alias{set_page}
\alias{set_page.pivot_table}
\title{Set page information to a pivot_table object}
\usage{
set_page(pt, row = 0, col = 0, page = "")

\method{set_page}{pivot_table}(pt, row = 0, col = 0, page = "")
}
\arguments{
\item{pt}{A pivot_table object.}

\item{row, col}{A cell (row and column number), page information included in
the table.}

\item{page}{A string, page information.}
}
\value{
A pivot_table object.
}
\description{
Define the page information associated with a pivot table. Previously
existing information is replaced by new information.
}
\details{
Each pivot table implements a report. The pivot table page represents the
context of that report. It is useful when we work with several pivot tables
with the same structure: for example, the page can allow us to differentiate
their origin, date or author. This information is often included in the file
name, sheet name, or a cell attached to the pivot table.
}
\examples{
library(tidyr)

pt <- pt_m4 \%>\% set_page(1, 1)

pt <- pt_m4 \%>\% set_page(page = "M4")

}
\keyword{internal}
