% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_table_view_table_attr.R
\name{view_table_attr}
\alias{view_table_attr}
\alias{view_table_attr.pivot_table}
\title{View table and attributes}
\usage{
view_table_attr(pt)

\method{view_table_attr}{pivot_table}(pt)
}
\arguments{
\item{pt}{A pivot_table object.}
}
\value{
A pivot_table object.
}
\description{
Displays the table and attributes of the object.
}
\keyword{internal}
